+++ apps/meteor/client/views/root/MainLayout/LoginPage.tsx
@@ -4,13 +4,13 @@ import RegistrationRoute from '@rocket.chat/web-ui-registration';
 import type { ReactElement, ReactNode } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import { useIframeLogin } from './useIframeLogin';
 import LoggedOutBanner from '../../../components/deviceManagement/LoggedOutBanner';
+import { useIframe } from '../../../hooks/iframe/useIframe';
 
 const LoginPage = ({ defaultRoute, children }: { defaultRoute?: LoginRoutes; children?: ReactNode }): ReactElement => {
 	const { t } = useTranslation();
 	const showForcedLogoutBanner = useSession('force_logout') as boolean | undefined;
-	const iframeLoginUrl = useIframeLogin();
+	const { iframeLoginUrl } = useIframe();
 
 	if (iframeLoginUrl) {
 		return <iframe title={t('Login')} src={iframeLoginUrl} style={{ height: '100%', width: '100%' }} />;
